requirements.txt
boto3==1.5.25
slackclient==2.5.0